+++ app/src/main/java/com/babylon/wallet/android/presentation/accessfactorsources/AccessFactorSourceNavGraph.kt
@@ -1,4 +1,4 @@
-package com.babylon.wallet.android.presentation.accessfactorsource
+package com.babylon.wallet.android.presentation.accessfactorsources
 
 import androidx.compose.ui.window.DialogProperties
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -6,18 +6,18 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.dialog
 
-fun NavController.accessFactorSourceBottomSheet() {
+fun NavController.accessFactorSources() {
     navigate("access_factor_source_bottom_sheet")
 }
 
-fun NavGraphBuilder.accessFactorSourceBottomSheet(
+fun NavGraphBuilder.accessFactorSources(
     onDismiss: () -> Unit
 ) {
     dialog(
         route = "access_factor_source_bottom_sheet",
         dialogProperties = DialogProperties(usePlatformDefaultWidth = false)
     ) {
-        AccessFactorSourceBottomSheet(
+        AccessFactorSourcesDialog(
             viewModel = hiltViewModel(),
             onDismiss = onDismiss
         )
