+++ app/src/main/java/com/babylon/wallet/android/presentation/accessfactorsources/AccessFactorSourcesViewModel.kt
@@ -1,4 +1,4 @@
-package com.babylon.wallet.android.presentation.accessfactorsource
+package com.babylon.wallet.android.presentation.accessfactorsources
 
 import com.babylon.wallet.android.presentation.common.StateViewModel
 import com.babylon.wallet.android.presentation.common.UiState
@@ -6,7 +6,7 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class AccessFactorSourceViewModel @Inject constructor() : StateViewModel<AccessFactorSourceViewModel.AccessFactorSourceUiState>() {
+class AccessFactorSourcesViewModel @Inject constructor() : StateViewModel<AccessFactorSourcesViewModel.AccessFactorSourceUiState>() {
 
     override fun initialState(): AccessFactorSourceUiState = AccessFactorSourceUiState()
