+++ app/src/main/java/com/babylon/wallet/android/presentation/navigation/NavigationHost.kt
@@ -10,7 +10,7 @@ import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
-import com.babylon.wallet.android.presentation.accessfactorsource.accessFactorSourceBottomSheet
+import com.babylon.wallet.android.presentation.accessfactorsources.accessFactorSources
 import com.babylon.wallet.android.domain.model.TransferableAsset
 import com.babylon.wallet.android.domain.model.resources.XrdResource
 import com.babylon.wallet.android.presentation.account.AccountScreen
@@ -241,7 +241,7 @@ fun NavigationHost(
                 }
             )
         }
-        accessFactorSourceBottomSheet(
+        accessFactorSources(
             onDismiss = {
                 navController.popBackStack()
             }
